add conv2doptions
