+++ example/mnist_cnn.go
@@ -1,0 +1,9 @@
+package main
+
+import "github.com/kuroko1t/gotorch/go"
+import "fmt"
+
+func main() {
+	model := torch.ModelInit()
+
+}
